// src/sidebars/types.ts
type Expand<T extends {[x: string]: unknown}> = {[P in keyof T]: T[P]};

export type SidebarItemBase = {
  className?: string;
  customProps?: {[key: string]: unknown};
};

export type SidebarItemDoc = SidebarItemBase & {
  type: 'tutorial' | 'ref';
  label?: string;
  id: string;
  /**
   * This is an internal marker. Items with labels defined in the config needs
   * to be translated with JSON
   */
  translatable?: true;
};

export type SidebarItemHtml = SidebarItemBase & {
  type: 'html';
  value: string;
  defaultStyle?: boolean;
};

export type SidebarItemLink = SidebarItemBase & {
  type: 'link';
  href: string;
  label: string;
  autoAddBaseUrl?: boolean;
  description?: string;
};

export type SidebarItemAutogenerated = SidebarItemBase & {
  type: 'autogenerated';
  dirName: string;
};

type SidebarItemCategoryBase = SidebarItemBase & {
  type: 'category';
  label: string;
  collapsed: boolean;
  collapsible: boolean;
  description?: string;
};

export type SidebarItemCategoryLinkDoc = {type: 'tutorial'; id: string};

export type SidebarItemCategoryLinkGeneratedIndexConfig = {
  type: 'generated-index';
  slug?: string;
  title?: string;
  description?: string;
  image?: string;
  keywords?: string | readonly string[];
};
export type SidebarItemCategoryLinkGeneratedIndex = {
  type: 'generated-index';
  slug: string;
  permalink: string;
  title?: string;
  description?: string;
  image?: string;
  keywords?: string | readonly string[];
};

export type SidebarItemCategoryLinkConfig =
  | SidebarItemCategoryLinkDoc
  | SidebarItemCategoryLinkGeneratedIndexConfig;

export type SidebarItemCategoryLink =
  | SidebarItemCategoryLinkDoc
  | SidebarItemCategoryLinkGeneratedIndex;
// ...
export type SidebarItemCategory = Expand<
  SidebarItemCategoryBase & {
    items: SidebarItem[];
    link?: SidebarItemCategoryLink;
  }
>;

export type SidebarItemCategoryWithLink = Required<SidebarItemCategory, 'link'>;

export type SidebarItemCategoryWithGeneratedIndex =
  SidebarItemCategoryWithLink & {link: SidebarItemCategoryLinkGeneratedIndex};

export type SidebarItem =
  | SidebarItemDoc
  | SidebarItemHtml
  | SidebarItemLink
  | SidebarItemCategory;